requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-modeltranslation==0.19.16
django-phonenumber-field==8.1.0
django-modeltranslation==0.19.16
pillow==11.3.0
drf-yasg==1.21.10
django-allauth==65.10.0
djangorestframework_simplejwt==5.5.1
phonenumbers==9.0.10
phonenumberslite==9.0.10
requests==2.32.4
cryptography==45.0.6